# profile/pr1-scaffold/openapi.yaml
openapi: 3.1.0
info:
  title: UbiquityOS Sprint Dashboard API (PR-1 Scaffold)
  version: 0.1.0
servers:
  - url: https://api.example.com
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    AuthSession:
      type: object
      description: Session metadata returned after OAuth callback. Access token is stored server-side and session is tracked via HttpOnly cookie.
      required: [userId, expiresAt]
      properties:
        userId:
          type: string
        expiresAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    Organization:
      type: object
      required: [id, login, name]
      properties:
        id:
          type: string
        login:
          type: string
        name:
          type: string

    JobAck:
      type: object
      required: [jobId, status]
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [queued]

    IngestionJobStatus:
      type: object
      required: [jobId, status, progress]
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        progress:
          type: number
          minimum: 0
          maximum: 100
        message:
          type: string
security:
  - cookieAuth: []
paths:
  /auth/github/start:
    get:
      security: []
      summary: Start GitHub OAuth login
      responses:
        '302':
          description: Redirect to GitHub authorization page

  /auth/github/callback:
    get:
      security: []
      summary: OAuth callback endpoint
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Auth successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '400':
          description: Missing required OAuth callback parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orgs:
    get:
      summary: List organizations available to current user
      responses:
        '200':
          description: Organization list
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  $ref: '#/components/schemas/Organization'
        '401':
          description: Missing or invalid session cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orgs/{orgId}/connect:
    post:
      summary: Connect organization and select repositories for ingestion
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repositories]
              properties:
                repositories:
                  type: array
                  minItems: 1
                  maxItems: 200
                  items:
                    type: string
      responses:
        '202':
          description: Connection accepted and ingestion enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAck'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid session cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /ingestion/jobs/{jobId}:
    get:
      summary: Poll ingestion job status
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJobStatus'
        '401':
          description: Missing or invalid session cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'